backend/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
python-dotenv>=0.20
Werkzeug>=2.0
python-docx==1.1.2
qrcode==7.4.2
Pillow==10.4.0
gunicorn>=20.0
PyOpenSSL>=22.0.0

# Database drivers (install one or both)
psycopg2-binary>=2.9.9
PyMySQL>=1.0

# SQLAlchemy
Flask-SQLAlchemy>=2.5
SQLAlchemy-serializer>=1.4.1
cryptography>=3.4.8
